// src/__generated__/deleteProduct.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

import { DeleteProductDto } from "./globalTypes";

// ====================================================
// GraphQL mutation operation: deleteProduct
// ====================================================

export interface deleteProduct_deleteProduct {
  __typename: "CommonOutput";
  ok: boolean | null;
  error: string | null;
}

export interface deleteProduct {
  deleteProduct: deleteProduct_deleteProduct;
}

export interface deleteProductVariables {
  input: DeleteProductDto;
}